Allow completions for vars in '.' namespace
Modify the regex for recognizing potential var names to allow for
namespaces starting with the '.' character. It is quite a popular choice
to put utility functions in the '.' namespace so that they are globally
available; for examples see [1] and [2]. Even though this may not be
officially supported, it is a good idea to support completions for this
namespace.

[1]: https://github.com/zcaudate/vinyasa
[2]: https://github.com/palletops/lein-shorthand